split akka-streams and doobie utils from ContractsFetch to new fetch-contracts library
