+++ ledger-service/fetch-contracts/BUILD.bazel
@@ -1,0 +1,196 @@
+# Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load(
+    "//bazel_tools:scala.bzl",
+    "da_scala_benchmark_jmh",
+    "da_scala_binary",
+    "da_scala_library",
+    "da_scala_test",
+    "lf_scalacopts",
+    "silencer_plugin",
+)
+load("//rules_daml:daml.bzl", "daml_compile")
+load("@os_info//:os_info.bzl", "is_windows")
+load("@scala_version//:index.bzl", "scala_version_suffix")
+
+hj_scalacopts = lf_scalacopts + [
+    "-P:wartremover:traverser:org.wartremover.warts.NonUnitStatements",
+]
+
+[
+    da_scala_library(
+        name = "http-json-{}".format(edition),
+        srcs = glob(["src/main/scala/**/*.scala"]),
+        plugins = [
+            "@maven//:org_typelevel_kind_projector_{}".format(scala_version_suffix),
+            silencer_plugin,
+        ],
+        scala_deps = [
+            "@maven//:com_chuusai_shapeless",
+            "@maven//:com_github_scopt_scopt",
+            "@maven//:com_typesafe_akka_akka_http",
+            "@maven//:com_typesafe_akka_akka_http_core",
+            "@maven//:io_spray_spray_json",
+            "@maven//:org_scala_lang_modules_scala_collection_compat",
+            "@maven//:org_scalaz_scalaz_core",
+            "@maven//:org_tpolecat_doobie_core",
+            "@maven//:org_tpolecat_doobie_free",
+            "@maven//:org_tpolecat_doobie_postgres",
+            "@maven//:org_typelevel_cats_core",
+            "@maven//:org_typelevel_cats_effect",
+            "@maven//:org_typelevel_cats_free",
+            "@maven//:org_typelevel_cats_kernel",
+        ],
+        scalacopts = hj_scalacopts,
+        tags = ["maven_coordinates=com.daml:http-json:__VERSION__"],
+        visibility = ["//visibility:public"],
+        runtime_deps = [
+            "@maven//:ch_qos_logback_logback_classic",
+            "@maven//:org_codehaus_janino_janino",
+        ],
+        deps = [
+            "//runtime-components/jdbc-drivers:jdbc-drivers-{}".format(edition),
+            "//daml-lf/data",
+            "//daml-lf/interface",
+            "//daml-lf/transaction",
+            "//language-support/scala/bindings-akka",
+            "//ledger-api/rs-grpc-bridge",
+            "//ledger-service/http-json-cli:{}".format(edition),
+            "//ledger-service/http-json-ledger-client:{}".format(edition),
+            "//ledger-service/cli-opts",
+            "//ledger-service/db-backend",
+            "//ledger-service/jwt",
+            "//ledger-service/lf-value-json",
+            "//ledger-service/utils",
+            "//ledger/ledger-api-auth",
+            "//ledger/ledger-api-common",
+            "//ledger/ledger-resources",
+            "//ledger/metrics",
+            "//libs-scala/concurrent",
+            "//libs-scala/contextualized-logging",
+            "//libs-scala/db-utils",
+            "//libs-scala/doobie-slf4j",
+            "//libs-scala/logging-entries",
+            "//libs-scala/ports",
+            "//libs-scala/resources",
+            "//libs-scala/scala-utils",
+            "@maven//:io_dropwizard_metrics_metrics_core",
+        ],
+    )
+    for edition in [
+        "ce",
+        "ee",
+    ]
+]
+
+json_scala_deps = [
+    "@maven//:com_chuusai_shapeless",
+    "@maven//:com_github_scopt_scopt",
+    "@maven//:com_typesafe_akka_akka_http",
+    "@maven//:com_typesafe_akka_akka_http_core",
+    "@maven//:com_typesafe_akka_akka_http_spray_json",
+    "@maven//:com_typesafe_akka_akka_slf4j",
+    "@maven//:com_typesafe_scala_logging_scala_logging",
+    "@maven//:io_spray_spray_json",
+    "@maven//:org_scala_lang_modules_scala_collection_compat",
+    "@maven//:org_scalaz_scalaz_core",
+    "@maven//:org_tpolecat_doobie_core",
+    "@maven//:org_tpolecat_doobie_free",
+    "@maven//:org_tpolecat_doobie_postgres",
+    "@maven//:org_typelevel_cats_core",
+    "@maven//:org_typelevel_cats_effect",
+    "@maven//:org_typelevel_cats_free",
+    "@maven//:org_typelevel_cats_kernel",
+]
+
+json_deps_shared = [
+    "//daml-lf/data",
+    "//daml-lf/interface",
+    "//daml-lf/transaction",
+    "//language-support/scala/bindings-akka",
+    "//ledger/metrics",
+    "//ledger-api/rs-grpc-bridge",
+    "//ledger-service/db-backend",
+    "//ledger-service/jwt",
+    "//ledger-service/lf-value-json",
+    "//ledger-service/utils",
+    "//ledger/ledger-api-auth",
+    "//ledger/ledger-api-common",
+    "//libs-scala/db-utils",
+    "//libs-scala/auth-utils",
+    "//libs-scala/ports",
+    "//libs-scala/scala-utils",
+    "@maven//:ch_qos_logback_logback_classic",
+    "@maven//:io_dropwizard_metrics_metrics_core",
+]
+
+json_deps = {
+    "ce": json_deps_shared + [
+        ":http-json-ce",
+        "//ledger-service/http-json-cli:ce",
+        "//ledger-service/http-json-ledger-client:ce",
+    ],
+    "ee": json_deps_shared + [
+        ":http-json-ee",
+        "//ledger-service/http-json-cli:ee",
+        "//ledger-service/http-json-ledger-client:ee",
+        "@maven//:com_oracle_database_jdbc_ojdbc8",
+    ],
+}
+
+[
+    da_scala_test(
+        name = "tests-{}".format(edition),
+        size = "medium",
+        srcs = glob(["src/test/scala/**/*.scala"]),
+        plugins = [
+            "@maven//:org_typelevel_kind_projector_{}".format(scala_version_suffix),
+            silencer_plugin,
+        ],
+        scala_deps = [
+            "@maven//:com_chuusai_shapeless",
+            "@maven//:com_typesafe_akka_akka_http_core",
+            "@maven//:io_spray_spray_json",
+            "@maven//:org_scala_lang_modules_scala_collection_compat",
+            "@maven//:org_scalacheck_scalacheck",
+            "@maven//:org_scalatest_scalatest_core",
+            "@maven//:org_scalatest_scalatest_matchers_core",
+            "@maven//:org_scalatest_scalatest_shouldmatchers",
+            "@maven//:org_scalatest_scalatest_wordspec",
+            "@maven//:org_scalatestplus_scalacheck_1_15",
+            "@maven//:org_scalaz_scalaz_core",
+            "@maven//:org_scalaz_scalaz_scalacheck_binding",
+            "@maven//:org_tpolecat_doobie_core",
+            "@maven//:org_tpolecat_doobie_free",
+            "@maven//:org_typelevel_cats_core",
+            "@maven//:org_typelevel_cats_effect",
+            "@maven//:org_typelevel_cats_free",
+            "@maven//:org_typelevel_cats_kernel",
+        ],
+        scalacopts = hj_scalacopts,
+        deps = [
+            ":http-json-{}".format(edition),
+            "//daml-lf/data",
+            "//daml-lf/interface",
+            "//daml-lf/transaction",
+            "//daml-lf/transaction-test-lib",
+            "//language-support/scala/bindings-akka",
+            "//ledger-service/db-backend",
+            "//ledger/metrics",
+            "//ledger-service/http-json-cli:{}".format(edition),
+            "//ledger-service/cli-opts",
+            "//ledger-service/lf-value-json",
+            "//ledger/ledger-api-common",
+            "//libs-scala/db-utils",
+            "//libs-scala/scala-utils",
+            "//libs-scala/scalatest-utils",
+            "@maven//:io_dropwizard_metrics_metrics_core",
+            "@maven//:org_scalatest_scalatest_compatible",
+        ],
+    )
+    for edition in [
+        "ce",
+        "ee",
+    ]
+]
